Change purgeban to use custom clean logic
This migrates the purgeban command away from Discord's native purgeban to our custom logic.

Discord's native purgeban does not leave us with any evidence or context of what messages were deleted. So when mods reference the infraction at  a later date they are lacking information.

Instead, we use our custom clean cog to delete all messages from the user in question for the last hour, and automatically append the link to the clean log to the infraction reason.

.